# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/Gearing/NewGearRatioNeeded/{newTireDiamIn}/{oldTireDiamIn}/{axleRatio}": {
      "get": {
        "tags": [
          "Gearing"
        ],
        "description": "Determine the new gear ratio needed when moving to a different  tire size to keep the overall ratio the same.Also provides the effective gear  ratio with the current gearing and new tire size",
        "parameters": [
          {
            "name": "newTireDiamIn",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Diameter in inches of new tire"
          },
          {
            "name": "oldTireDiamIn",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Diameter in inches of old tire"
          },
          {
            "name": "axleRatio",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Ring & pinion ratio before the colon, ex. 3.73"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Gearing/EngineRpmAtSpeed/{axleRatio}/{vehicleSpeedMph}/{transRatio}/{tireDiamIn}": {
      "get": {
        "tags": [
          "Gearing"
        ],
        "description": "Determine the engine RPM at a given speed for the combination of the  axle ratio, transmission gear and tire diameter",
        "parameters": [
          {
            "name": "axleRatio",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Ring & pinion ratio before the colon, ex. 3.73"
          },
          {
            "name": "vehicleSpeedMph",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Vehicle speed in MPH"
          },
          {
            "name": "transRatio",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Transmission ratio before the colon, ex. 2.54"
          },
          {
            "name": "tireDiamIn",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Diameter in inches of tire"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Gearing/VehicleSpeedAtEngineRpm/{engineRpm}/{tireDiamIn}/{transRatio}/{auxRatio}/{tCaseRatio}/{axleRatio}": {
      "get": {
        "tags": [
          "Gearing"
        ],
        "description": "Determine the vehicle speed at a given engine RPM for the  combination of the tire diameter, transmission gear,  overdrive (if equipped, else 1, can also be used to figure a second transfer case),  transfer case (high or low gear as needed), and axle ratio",
        "parameters": [
          {
            "name": "engineRpm",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "RPM of the engine"
          },
          {
            "name": "tireDiamIn",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Diameter in inches of tire"
          },
          {
            "name": "transRatio",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Transmission ratio before the colon, ex. 2.54"
          },
          {
            "name": "auxRatio",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "auxRatio Overdrive or second transfer case ratio before the colon,  ex. 4.0. or 1 if not equipped"
          },
          {
            "name": "tCaseRatio",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Transfer case ratio before the colon, ex. 2.72"
          },
          {
            "name": "axleRatio",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Ring & pinion ratio before the colon, ex. 3.73"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Gearing/CrawlRatio/{axleRatio}/{lowTCaseRatio}/{transLowGearRatio}/{auxRatio}": {
      "get": {
        "tags": [
          "Gearing"
        ],
        "description": "Determine the crawl ratio for the provided parameters",
        "parameters": [
          {
            "name": "axleRatio",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Ring & pinion ratio before the colon, ex. 3.73"
          },
          {
            "name": "lowTCaseRatio",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Transfer case ratio before the colon, ex. 2.72"
          },
          {
            "name": "transLowGearRatio",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Transmission ratio before the colon, ex. 2.54"
          },
          {
            "name": "auxRatio",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "auxRatio Overdrive or second transfer case ratio before the colon,  ex. 4.0. or 1 if not equipped"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Gearing/RotationVelocity/{axleRatio}/{speedMph}/{tireDiamIn}": {
      "get": {
        "tags": [
          "Gearing"
        ],
        "description": "Determine the rotation velocity which is how fast  an object rotates or revolves relative to another point",
        "parameters": [
          {
            "name": "axleRatio",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Ring & pinion ratio before the colon, ex. 3.73"
          },
          {
            "name": "speedMph",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Vehicle speed in MPH"
          },
          {
            "name": "tireDiamIn",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Diameter in inches of tire"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Gearing/CalculateGearRatio/{ringGearTeeth}/{pinionGearTeeth}": {
      "get": {
        "tags": [
          "Gearing"
        ],
        "description": "Determine the gear ratio by counting teeth on the ring and pinion",
        "parameters": [
          {
            "name": "ringGearTeeth",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Number of teeth on the ring gear"
          },
          {
            "name": "pinionGearTeeth",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Number of teeth on the pinion gear"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Gearing/ActualSpeed/{newTireDiamIn}/{oldTireDiamIn}/{speedoMph}": {
      "get": {
        "tags": [
          "Gearing"
        ],
        "description": "Determine actual speed when changing tire sizes and not regearing.  The speedometer is still calibrated at factory specs and this will show what  speed you are actually running",
        "parameters": [
          {
            "name": "newTireDiamIn",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Diameter in inches of new tire"
          },
          {
            "name": "oldTireDiamIn",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Diameter in inches of old tire"
          },
          {
            "name": "speedoMph",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Vehicle speed in MPH as read on the speedometer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}